building songs screen

+++ app/src/main/java/com/videomaker/aimusic/domain/model/VideoTemplate.kt
@@ -16,7 +16,7 @@ data class VideoTemplate(
     val name: String,
     val description: String,
     val thumbnailUrl: String = "",
-    val songId: Int,
+    val songId: Long,
     val effectSetId: String,
     val aspectRatio: String = "9:16",
     val tags: List<String> = emptyList(),
